third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-create-basics.html

<!DOCTYPE html>
<title>Credential Manager: create() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
promise_test(function(t) {
    var credential_data = {
        id: 'id',
        password: 'pencil',
    };

    return navigator.credentials.create({password: credential_data})
        .then(function(credential) {
            assert_equals(credential.idName, 'username');
            assert_equals(credential.passwordName, 'password');
            assert_equals(credential.additionalData, null);
        });
}, "navigator.credentials.create() with valid PasswordCredentialData");

promise_test(function(t) {
    var f = document.createElement('form');
    f.innerHTML = "<input type='text' name='theId' value='musterman' autocomplete='username'>"
        + "<input type='text' name='thePassword' value='sekrit' autocomplete='current-password'>"
        + "<input type='text' name='theIcon' value='https://example.com/photo' autocomplete='photo'>"
        + "<input type='text' name='theExtraField' value='extra'>"
        + "<input type='text' name='theName' value='friendly name' autocomplete='name'>";

    return navigator.credentials.create({password: f})
        .then(function(credential) {
            assert_equals(credential.idName, 'theId');
            assert_equals(credential.passwordName, 'thePassword');

            assert_equals(credential.additionalData.get('theId'), 'musterman');
            assert_equals(credential.additionalData.get('thePassword'), 'sekrit');
            assert_equals(credential.additionalData.get('theIcon'),
                    'https://example.com/photo');
            assert_equals(credential.additionalData.get('theName'), 'friendly name');
            assert_equals(credential.additionalData.get('theExtraField'), 'extra');
        });
}, "navigator.credentials.create() with valid HTMLFormElement");

add_completion_callback(() => {
  if (window.testRunner)
    window.testRunner.clearMockAuthenticatorResponse();
});

promise_test(() => {
  var raw_id = new Uint8Array(new TextEncoder("utf-8").encode("raw_id"));
  var id = btoa("raw_id");
  console.log(id);
  var client_data_json = new Uint8Array(new TextEncoder("utf-8").encode("client_data_json"));
  var attestation_object = new Uint8Array(new TextEncoder("utf-8").encode("attestation_object"));
  var authenticator_data = { attestation_object }

  if (window.testRunner)
    testRunner.setMockAuthenticatorResponse(id, raw_id, client_data_json, attestation_object, new Uint8Array(0), new Uint8Array(0));

  var publicKey = {
    challenge: new TextEncoder().encode("climb a mountain"),
    // Relying Party:
    rp: {
      id: "1098237235409872",
      name: "Acme"
    },
  
    // User:
    user: {
      id: "1098237235409872",
      name: "john.p.smith@example.com",
      displayName: "John P. Smith",
      icon: "https://pics.acme.com/00/p/aBjjjpqPb.png"
    },
  
    // This Relying Party will accept either an ES256 or RS256 credential, but
    // prefers an ES256 credential.
    parameters: [
      {
        type: "public-key",
        algorithm: "ES256",
      },
      {
        type: "public-key",
        algorithm: "RS256",
      },
    ],

    timeout: 60000,  // 1 minute
    excludeList: [], // No excludeList
    extensions: {"webauthn.location": true}  // Include location information
                                             // in attestation
  };

  return navigator.credentials.create({publicKey}).then(r => {
    console.log(r);
    console.log(r.__proto__);
    console.log("Logging things");
    console.log(r.id);
    console.log(r.raw_id);
    console.log(r.client_data_json);
    assert_equals(r.id, id, "id");    
    assert_equals(r.raw_id, raw_id, "name");
    assert_equals(r.client_data_json, client_data_json, "client_data_json");
    assert_equals(r.attestation_object.authenticator_data, authenticator_data, "authenticator_data");
    assert_not_exists(r.attestation_object, attestation_data, "attestation_data");
    assert_not_exists(r.attestation_object, signature, "signature");
  });
}, "Verify that the mock returns the values we give it.");

</script>
